% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Clean Up Unneeded Files from Working Directory}
\usage{
cleanup(converted_audio = FALSE, transcripts = FALSE, textgrids = FALSE)
}
\arguments{
\item{converted_audio}{Remove converted audio files? (default: FALSE)}

\item{transcripts}{Remove transcriptions exported as .TXT? (default: FALSE)}

\item{textgrids}{Remove .TextGrid files? (default: FALSE)}
}
\description{
This function provides a simple interface for quickly removing any of the output produced by scrivenR.
}
\examples{
cleanup(converted audio = FALSE, transcripts = FALSE, textgrids = FALSE)
}
\keyword{ASR,}
\keyword{batch-processing,}
\keyword{file}
\keyword{removal}
\keyword{transcription,}
\keyword{whisper,}
