% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_audio.R
\name{convert_audio}
\alias{convert_audio}
\title{Convert Audio File to 16-bit Mono .WAV}
\usage{
convert_audio(x)
}
\arguments{
\item{x}{Output of list.files() in working directory containing videos}
}
\description{
This function takes a list of files, filters by common audio file extensions, and then runs an ffmpeg command to convert the audio and write output to the working directory.
}
\examples{
convert_audio(list.files())
}
\keyword{audio}
\keyword{conversion}
