% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_audio.R
\name{extract_audio}
\alias{extract_audio}
\title{Extract Audio File from Video}
\usage{
extract_audio(x)
}
\arguments{
\item{x}{Output of list.files() in working directory containing videos}
}
\description{
This function takes a list of files, filters by common video file extensions, and then runs an ffmpeg command to extract the audio and write output to the working directory.
}
\examples{
extract_audio(list.files())
}
\keyword{audio}
\keyword{conversion,}
\keyword{extraction}
