% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_audio.R
\name{transcribe_audio}
\alias{transcribe_audio}
\title{Transcribe 16-bit Mono .WAV file}
\usage{
transcribe_audio(
  x,
  model_path,
  include_timing = FALSE,
  internal_convert = TRUE,
  write_textgrids = FALSE,
  print_trace = FALSE,
  prompt = "",
  n_threads = 1
)
}
\arguments{
\item{x}{Output of list.files() in working directory containing audio}

\item{model_path}{A character string of the path for your local Whisper acoustic model; this is required}

\item{include_timing}{Option to include timestamps for each line of text, default=TRUE}

\item{internal_convert}{Option to retain naming format for conversions made with this package, default=TRUE. Setting to FALSE will point the function at any .WAVs in the working directory.}

\item{write_textgrids}{Option to print time-aligned .TextGrid files for use in Praat; FALSE by default}

\item{print_trace}{Option to print the on-going transcript in the console, default = FALSE}

\item{prompt}{Option to provide an initial prompt to the model}

\item{n_threads}{The number of CPU threads to be used in parallel processing. Default value is 1, i.e. single-thread processing.}
}
\description{
This function takes a list of file names for 16-bit mono WAVs and applies batch processing of automatic transcription using Whisper.
}
\examples{
transcribe_audio(list.files(), model_path = mp, include_timing=TRUE, internal_convert=TRUE, write_textgrids = FALSE, print_trace = FALSE, n_threads = FALSE)
}
\keyword{ASR,}
\keyword{batch-processing}
\keyword{transcription,}
\keyword{whisper,}
