% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_linux.R
\name{transcribe_linux}
\alias{transcribe_linux}
\title{Transcribe 16-bit Mono .WAV file}
\usage{
transcribe_linux(
  x,
  model_path,
  include_timing = FALSE,
  internal_convert = TRUE,
  write_textgrids = FALSE,
  n_threads = 1
)
}
\arguments{
\item{x}{Output of list.files() in working directory containing audio}

\item{model_path}{A character string of the path for your local Whisper acoustic model; this is required}

\item{include_timing}{Option to include timestamps for each line of text, default=TRUE}

\item{internal_convert}{Option to retain naming format for conversions made with this package, default=TRUE. Setting to FALSE will point the function at any .WAVs in the working directory.}

\item{write_textgrids}{Option to print time-aligned .TextGrid files for use in Praat; FALSE by default}

\item{n_threads}{Number of CPU threads to use for parallel processing (default = 1). Use parallel::detectCores() if you are not sure how many you can use.}
}
\description{
This function takes a list of file names for 16-bit mono WAVs and applies batch processing of automatic transcription using Whisper.
}
\examples{
transcribe_linux(list.files(), model_path = mp, include_timing=TRUE, internal_convert=TRUE, write_textgrids = FALSE, n_threads = 1)
}
\keyword{ASR,}
\keyword{batch-processing}
\keyword{transcription,}
\keyword{whisper,}
